requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
appdirs==1.4.4
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.2
backoff==2.2.1
backports.zoneinfo==0.2.1
bcrypt==3.2.0
black==23.7.0
blinker==1.6.2
boto3==1.28.13
botocore==1.31.13
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.5
cloudpickle==2.0.0
constantly==15.1.0
cryptography==39.0.1
daphne==4.0.0
db-dtypes==1.1.1
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==2.0.0
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.54.0
django-cors-headers==4.1.0
django-ses==3.5.0
django-silk==5.0.3
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
filelock==3.12.2
frozenlist==1.4.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
gunicorn==21.2.0
gprof2dot==2022.7.29
grpcio==1.57.0
grpcio-status==1.57.0
hyperlink==21.0.0
identify==2.5.24
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
incremental==22.10.0
jmespath==1.0.1
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.8.0
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.8
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
pip==23.1.2
platformdirs==3.8.1
pre-commit==3.3.3
proto-plus==1.22.3
protobuf==4.24.0
psycopg==3.1.9
psycopg2==2.9.3
psycopg2-binary==2.9.6
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.18.0
pydantic==2.0.2
pydantic_core==2.1.2
PyJWT==2.7.0
pyOpenSSL==23.2.0
PyPika==0.48.9
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
service-identity==18.1.0
setuptools==67.8.0
six==1.16.0
sly==0.5
snowflake-connector-python==3.0.4
snowflake-snowpark-python==1.5.1
sortedcontainers==2.4.0
sql-metadata==2.8.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.65.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.23.1
wheel==0.38.4
xlrd==2.0.1
yarl==1.9.2
zipp==3.16.2
zope.interface==5.4.0
